Add export to Frab functionality
For scheduling the FOSS4G 2016, Frab [1] will be used. Hence some
export functionality is needed.

The Frab that is used for the FOSS4G is also used for the FOSSGIS, hence
we use an offset for the unique IDs so that we don't end up with any
clashes.

To get the data into Frab, use these end points:

    /proposals/frab/events
    /proposals/frab/event_people
    /proposals/frab/people

like:

    curl -X GET 'http://localhost:8000/proposals/frab/events?ids=180,181,241,125,500' > frab/tmp/frab_export/events.yaml
    curl -X GET 'http://localhost:8000/proposals/frab/event_people?ids=180,181,241,125,500'> frab/tmp/frab_export/event_people.yaml
    curl -X GET 'http://localhost:8000/proposals/frab/people'> frab/frab/tmp/frab_export/people.yaml

It will export the requested proposals and all speakers from the system,
except the ones that haven't given a name.

[1]: https://github.com/frab/frab